[dynamo] Support BUILTIN_MATCH serialization.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/157016

Serialize BUILTIN_MATCH since they are all stored in __builtin__ dict.

Also fixed an issue that the wrong global scope is passed to CheckFunctionManager while loading guards. Previously we can always reuse the compile-time global scope for evaluating guards because the compile-time and runtime global scope are always the same.

For precompile, we need to serialize the compile-time global scope for loading only. We need to point the CheckFunctionManager to the new global scope after loading is finished for evaluating guards.
ghstack-source-id: 293653150
@exported-using-ghexport

Differential Revision: [D77159313](https://our.internmc.facebook.com/intern/diff/D77159313/)