Add a flag "realized" in IRNode to enable tracking origin_nodes (#157423)
Summary:

This DIFF is to fix the following issue:

In python source code for CompiledFxGraph,the FX graph segment for the Triton kernel is broken. For example, the following function
    def fn(a, b, c):
        x = torch.nn.functional.linear(a, b)
        x = x.sin()
        x = x.t() + c
        return x
Inductor compiled this FX graph into two nodes: the first one is mm, the second one is a triton kernel for sin + transpose + add. The FX graph segment for the triton kernel is like the following:
Graph fragment:
%add : [num_users=1] = call_function[target=torch.ops.aten.add.Tensor](args = (%permute_1, %arg2_1), kwargs = {})
Basically only "add" node in the FX graph.


The root cause is function caffe2/torch/_inductor/utils.py:gather_origins does not detect the realized node correctly.

To fix this issue, a new flag "realized" is added into IRNode with initial value "False", when self.realized() is called, that flag is set. With this flag, inside function gather_origins, only need to check this flag to determine the node is a realized node or not.

Rollback Plan:

Differential Revision: D77638777